// packages/cms/src/admin/__generated__/app-config.ts
// Auto-generated by Blazing CMS — do not edit
export const projectName = "Blazing CMS Playground";
export const capabilities: import("@blazing-cms/types").ResolvedCapabilities = {
  "collections": {
    "date": {
      "versioning": false
    }
  },
  "features": {
    "content": true,
    "analytics": false,
    "media": true,
    "versioning": true,
    "workflow": true,
    "notifications": true,
    "rbac": true
  },
  "globals": {}
};